error handling methods

+++ Projects/CommandDB/src/commands/DeleteCommand.java
@@ -12,6 +12,9 @@ public class DeleteCommand implements Command<Arguments, Void> {
 
     @Override
     public Void execute(Arguments args) {
+        if (args.size() < 1)
+            throw new IllegalArgumentException("Not enough arguments. Usage: delete [...]");
+
         for (Object arg : args) {
             int id = (Integer)arg;
             this.database.removeEntity(id);
